+++ web/src/constants/routes.ts
@@ -1,7 +1,7 @@
 export enum Routes {
   HOME = '/',
-  REQUESTS = 'requests',
-  SETTINGS = 'settings',
+  REQUESTS = '/requests',
+  SETTINGS = '/settings',
   SIGN_IN = '/auth/sign-in',
   SIGN_UP = '/auth/sign-up'
 }
